# frontend/src/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crypto Radar API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/trades": {
      "get": {
        "tags": [
          "trades"
        ],
        "summary": "List Trade Journal",
        "operationId": "list_trade_journal_api_v1_trades_get",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(virtual|real)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "signal_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Signal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradeJournalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trades/virtual": {
      "get": {
        "tags": [
          "trades"
        ],
        "summary": "List Virtual Trades",
        "operationId": "list_virtual_trades_api_v1_trades_virtual_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "signal_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Signal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VirtualTradeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trades/real": {
      "get": {
        "tags": [
          "trades"
        ],
        "summary": "List Real Trades",
        "operationId": "list_real_trades_api_v1_trades_real_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "signal_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Signal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradeJournalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trades/real/import": {
      "post": {
        "tags": [
          "trades"
        ],
        "summary": "Import Real Trades",
        "operationId": "import_real_trades_api_v1_trades_real_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RealTradeImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RealTradeImportResult"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RealTradeImportNotReadyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trades/real/external-orders": {
      "get": {
        "tags": [
          "trades"
        ],
        "summary": "List External Exchange Orders",
        "operationId": "list_external_exchange_orders_api_v1_trades_real_external_orders_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo_user",
              "title": "User Id"
            }
          },
          {
            "name": "connection_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Connection Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExternalExchangeOrderResponse"
                  },
                  "title": "Response List External Exchange Orders Api V1 Trades Real External Orders Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trades/real/external-trades": {
      "get": {
        "tags": [
          "trades"
        ],
        "summary": "List External Exchange Trades",
        "operationId": "list_external_exchange_trades_api_v1_trades_real_external_trades_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo_user",
              "title": "User Id"
            }
          },
          {
            "name": "connection_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Connection Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExternalExchangeTradeResponse"
                  },
                  "title": "Response List External Exchange Trades Api V1 Trades Real External Trades Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/trades/virtual/{trade_id}": {
      "get": {
        "tags": [
          "trades"
        ],
        "summary": "Get Virtual Trade",
        "operationId": "get_virtual_trade_api_v1_trades_virtual__trade_id__get",
        "parameters": [
          {
            "name": "trade_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trade Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VirtualTrade"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trades/{trade_id}": {
      "get": {
        "tags": [
          "trades"
        ],
        "summary": "Get Trade Journal Entry",
        "operationId": "get_trade_journal_entry_api_v1_trades__trade_id__get",
        "parameters": [
          {
            "name": "trade_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trade Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradeJournalEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trades/virtual/{trade_id}/close": {
      "post": {
        "tags": [
          "trades"
        ],
        "summary": "Close Virtual Trade",
        "operationId": "close_virtual_trade_api_v1_trades_virtual__trade_id__close_post",
        "parameters": [
          {
            "name": "trade_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trade Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseVirtualTradeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VirtualTrade"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CloseVirtualTradeRequest": {
        "properties": {
          "exit_price": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Price"
          },
          "reason": {
            "type": "string",
            "enum": [
              "take_profit",
              "stop_loss",
              "manual_close",
              "cancelled"
            ],
            "title": "Reason",
            "default": "manual_close"
          }
        },
        "type": "object",
        "title": "CloseVirtualTradeRequest"
      },
...
      "ExternalExchangeOrderResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "connection_id": {
            "type": "string",
            "format": "uuid",
            "title": "Connection Id"
          },
          "exchange_order_id": {
            "type": "string",
            "title": "Exchange Order Id"
          },
          "pair_id": {
            "type": "string",
            "format": "uuid",
            "title": "Pair Id"
          },
          "exchange_code": {
            "type": "string",
            "title": "Exchange Code"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "side": {
            "type": "string",
            "title": "Side"
          },
          "order_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Type"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "price": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "created_exchange_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created Exchange At"
          },
          "updated_exchange_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated Exchange At"
          },
          "imported_at": {
            "type": "string",
            "format": "date-time",
            "title": "Imported At"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "connection_id",
          "exchange_order_id",
          "pair_id",
          "exchange_code",
          "symbol",
          "side",
          "order_type",
          "status",
          "quantity",
          "price",
          "created_exchange_at",
          "updated_exchange_at",
          "imported_at",
          "metadata"
        ],
        "title": "ExternalExchangeOrderResponse"
      },
      "ExternalExchangeTradeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "connection_id": {
            "type": "string",
            "format": "uuid",
            "title": "Connection Id"
          },
          "exchange_trade_id": {
            "type": "string",
            "title": "Exchange Trade Id"
          },
          "exchange_order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange Order Id"
          },
          "pair_id": {
            "type": "string",
            "format": "uuid",
            "title": "Pair Id"
          },
          "exchange_code": {
            "type": "string",
            "title": "Exchange Code"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "side": {
            "type": "string",
            "title": "Side"
          },
          "price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price"
          },
          "quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Quantity"
          },
          "fee_amount": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fee Amount"
          },
          "fee_asset_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fee Asset Id"
          },
          "fee_asset_symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fee Asset Symbol"
          },
          "traded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Traded At"
          },
          "imported_at": {
            "type": "string",
            "format": "date-time",
            "title": "Imported At"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "connection_id",
          "exchange_trade_id",
          "exchange_order_id",
          "pair_id",
          "exchange_code",
          "symbol",
          "side",
          "price",
          "quantity",
          "fee_amount",
          "fee_asset_id",
          "fee_asset_symbol",
          "traded_at",
          "imported_at",
          "metadata"
        ],
        "title": "ExternalExchangeTradeResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RealTradeImportNotReadyResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "not_implemented",
            "title": "Status",
            "default": "not_implemented"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "connection_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connection Id"
          },
          "connector_required": {
            "type": "boolean",
            "title": "Connector Required",
            "default": true
          },
          "normalized_targets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Normalized Targets"
          },
          "analytics_targets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Analytics Targets"
          },
          "raw_targets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Raw Targets"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "RealTradeImportNotReadyResponse"
      },
      "RealTradeImportRequest": {
        "properties": {
          "connection_id": {
            "type": "string",
            "format": "uuid",
            "title": "Connection Id"
          },
          "since": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Since"
          },
          "until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Until"
          },
          "symbols": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Symbols"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "connection_id"
        ],
        "title": "RealTradeImportRequest"
      },
      "RealTradeImportResult": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "dry_run"
            ],
            "title": "Status"
          },
          "connection_id": {
            "type": "string",
            "format": "uuid",
            "title": "Connection Id"
          },
          "external_exchange_orders_written": {
            "type": "integer",
            "title": "External Exchange Orders Written",
            "default": 0
          },
          "external_exchange_trades_written": {
            "type": "integer",
            "title": "External Exchange Trades Written",
            "default": 0
          },
          "analytics_events_written": {
            "type": "integer",
            "title": "Analytics Events Written",
            "default": 0
          },
          "raw_events_written": {
            "type": "integer",
            "title": "Raw Events Written",
            "default": 0
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "type": "string",
            "format": "date-time",
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "status",
          "connection_id",
          "started_at",
          "finished_at"
        ],
        "title": "RealTradeImportResult"
      },
...
      "TradeJournalResponse": {
        "properties": {
          "trades": {
            "items": {
              "$ref": "#/components/schemas/TradeJournalEntry"
            },
            "type": "array",
            "title": "Trades"
          },
          "account": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VirtualAccount"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "trades"
        ],
        "title": "TradeJournalResponse"
      },
...
      "VirtualAccount": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": "demo_user"
          },
          "starting_balance": {
            "type": "number",
            "title": "Starting Balance",
            "default": 100.0
          },
          "balance": {
            "type": "number",
            "title": "Balance",
            "default": 100.0
          },
          "equity": {
            "type": "number",
            "title": "Equity",
            "default": 100.0
          },
          "realized_pnl": {
            "type": "number",
            "title": "Realized Pnl",
            "default": 0.0
          },
          "unrealized_pnl": {
            "type": "number",
            "title": "Unrealized Pnl",
            "default": 0.0
          },
          "risk_per_trade": {
            "type": "number",
            "title": "Risk Per Trade",
            "default": 10.0
          },
          "risk_reward": {
            "type": "number",
            "title": "Risk Reward",
            "default": 3.0
          },
          "open_positions": {
            "type": "integer",
            "title": "Open Positions",
            "default": 0
          },
          "closed_trades": {
            "type": "integer",
            "title": "Closed Trades",
            "default": 0
          },
          "wins": {
            "type": "integer",
            "title": "Wins",
            "default": 0
          },
          "losses": {
            "type": "integer",
            "title": "Losses",
            "default": 0
          },
          "breakeven": {
            "type": "integer",
            "title": "Breakeven",
            "default": 0
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "updated_at"
        ],
        "title": "VirtualAccount"
      },
...
      "VirtualTradeResponse": {
        "properties": {
          "trades": {
            "items": {
              "$ref": "#/components/schemas/VirtualTrade"
            },
            "type": "array",
            "title": "Trades"
          }
        },
        "type": "object",
        "required": [
          "trades"
        ],
        "title": "VirtualTradeResponse"
      },
...
    }
  }
}